Complete exercise
We first moved the ! icon into a separate div so that it could be arranged next to the rest of the content, which would be placed in it's own container. We then included the close button into the same div as the header text so that we could more easily manipulate the two in relation to each other. Finally we justified the content and made the final changes to the text. We also gave the icon a flex-shrink value of 0 so that it would not get squished if the textbox were to be resized.